Refactoring methods to common.go
